requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
jupyter>=1.0.0
pytest>=9.0.0
python-pptx>=1.0.2
